refactor: isolate generic live publishing
